Refactor code according PSR-4. Optimize DB inserts with batch flush and EM clear
