+++ docker/php/Dockerfile
@@ -10,6 +10,7 @@ RUN apt-get update && apt-get install -y \
         && docker-php-ext-install pdo pdo_mysql mbstring zip
 
 COPY --from=composer:latest /usr/bin/composer /usr/bin/composer
+COPY docker/php/conf.d/custom.ini /usr/local/etc/php/conf.d/
 
 WORKDIR /var/www
